garmin_sync/requirements.txt
garminconnect>=0.3.0
requests>=2.28.0
python-dotenv>=1.0.0
pynacl>=1.5.0
